Updated/Fixed enemy rotation
Updated the way enemies move to the next point in the walk cycle. The enemy game object no longer rotates, instead there is an enemy vision container child that contains the enemy vision game object and rotates towards the next point. This change was implemented so the human sprites are not rotated when walking.